move pointing files to src
