+++ keyboards/ergohaven/k03/rules.mk
@@ -39,4 +39,4 @@ SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/src/oled/eh_oled.c
 SRC += oled/ergohaven_dark.c
 SRC += keyboards/ergohaven/hid.c
-SRC += keyboards/ergohaven/ergohaven_pointing.c
+SRC += keyboards/ergohaven/src/eh_pointing.c
