+++ keyboards/ergohaven/k03pro/rules.mk
@@ -44,7 +44,7 @@ SRC += keyboards/ergohaven/ergohaven_main.c
 SRC += keyboards/ergohaven/src/eh_settings.c
 SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
-SRC += keyboards/ergohaven/ergohaven_pointing.c
+SRC += keyboards/ergohaven/src/eh_pointing.c
 
 SRC += keyboards/ergohaven/src/display/lvgl_helpers.c
 SRC += keyboards/ergohaven/src/display/eh_display_modes.c
