+++ keyboards/ergohaven/macropad/rev3/rules.mk
@@ -50,5 +50,5 @@ SRC += keyboards/ergohaven/src/display/fonts/eh_font_montserrat_28.c
 SRC += keyboards/ergohaven/ergohaven_main.c
 SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
-SRC += keyboards/ergohaven/ergohaven_pointing.c
+SRC += keyboards/ergohaven/src/eh_pointing.c
 SRC += keyboards/ergohaven/src/eh_settings.c
