+++ keyboards/ergohaven/sm30/rules.mk
@@ -39,7 +39,7 @@ UNICODE_ENABLE = yes
 SRC += keyboards/ergohaven/ergohaven_main.c
 SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
-SRC += keyboards/ergohaven/ergohaven_pointing.c
+SRC += keyboards/ergohaven/src/eh_pointing.c
 SRC += keyboards/ergohaven/src/eh_settings.c
 
 SRC += keyboards/ergohaven/src/display/eh_display.c
